libc: remove duplicate types from linuxlike/Mod.kt + fix linux/Mod.kt imports — 760 → 647 errors

+++ src/commonMain/kotlin/io/github/kotlinmania/libc/unix/Mod.kt
@@ -205,8 +205,8 @@ public const val INADDR_LOOPBACK: InAddrT = 2130706433u
 public const val INADDR_ANY: InAddrT = 0u
 public const val INADDR_BROADCAST: InAddrT = 4294967295u
 public const val INADDR_NONE: InAddrT = 4294967295u
-public val IN6ADDR_LOOPBACK_INIT: In6Addr = In6Addr(s6Addr = ubyteArrayOf(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1))
-public val IN6ADDR_ANY_INIT: In6Addr = In6Addr(s6Addr = ubyteArrayOf(0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0))
+public val IN6ADDR_LOOPBACK_INIT: In6Addr = In6Addr(s6Addr = ubyteArrayOf(0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 1.toUByte()))
+public val IN6ADDR_ANY_INIT: In6Addr = In6Addr(s6Addr = ubyteArrayOf(0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte(), 0.toUByte()))
 public val ARPOP_REQUEST: UShort = (1).toUShort()
 public val ARPOP_REPLY: UShort = (2).toUShort()
 public const val ATF_COM: CInt = 0x02
